# src/main/java/es/taller/alex/api.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:

  /car/create:
    post:
      tags:
        - car-controller
      operationId: postCar
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarCreateDto'

  /car/read/{carId}:
    get:
      tags:
        - car-controller
      operationId: getCar
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarReadDto'

  /car/update/{carId}:
    put:
      tags:
        - car-controller
      operationId: updateCar
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/compnents/schemas/CarUpdateDto'

    /car/patch/{carId}:
      patch:
        tags:
          - car-controller
        operationId: patchCar
        parameters:
          - name: carId
            in: path
            required: true
            schema:
              type: integer
              format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '/#/components/schemas/CarPatchDto'

    /car/delete/{carId}:
      delete:
        tags:
          - car-controller
        operationId: deleteCar
        parameters:
          - name: carId
            in: path
            required: true
            schema:
              type: integer
              format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '/#/components/schemas/CarDto'

    /client/read/{DNI}:
      get:
        tags:
          - client-controller
        operationId: getClient
        parameters:
          - name: DNI
            in: path
            required: true
            schema:
              type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '/#/components/schemas/ClientReadDto'

    /client/create:
      post:
        tags:
          - client-controller
        operationId: postClient
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '/#/components/schemas/ClientCreateDto'

    /client/update/{DNI}:
      update:
        tags:
          - client-controller
        operationId: updateClient
        parameters:
          - name: DNI
            in: path
            required: true
            schema:
              type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '/#/components/schemas/ClientUpdateDto'

    /client/patch/{DNI}:
      patch:
        tags:
          - client-controller
        operationId: patchClient
        parameters:
          - name: DNI
            in: path
            required: true
            schema:
              type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '/#/components/schemas/ClientPatchDto'

    /client/delete/{DNI}:
      delete:
        tags:
          - client-controller
        operationId: deleteClient
        parameters:
          - name: DNI
            in: path
            required: true
            schema:
              type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '/#/components/schemas/ClientDto'

    /mechanic/read/{DNI}:
      get:
        tags:
          - mechanic-controller
        operationId: getMechanic
        parameters:
          - name: DNI
            in: path
            required: true
            schema:
              type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '/#/components/schemas/MechanicReadDto'

    /mechanic/create:
      post:
        tags:
          - mechanic-controller
        operationId: postMechanic
        responses:
          '200':
            description: OK
            content:
              application/json:
                schema:
                  $ref: '/#/components/schemas/MechanicCreateDto'

    /mechanic/update/{DNI}:
      update:
        tags:
          - mechanic-controller
        operationId: updateMechanic
        parameters:
          - name: DNI
            in: path
            required: true
            schema:
              type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '/#/components/schemas/MechanicUpdateDto'

    /mechanic/patch/{DNI}:
      patch:
        tags:
          - mechanic-controller
        operationId: patchMechanic
        parameters:
          - name: DNI
            in: path
            required: true
            schema:
              type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string

    /mechanic/delete/{DNI}:
      delete:
        tags:
          - mechanic-controller
        operationId: deleteMechanic
        parameters:
          - name: DNI
            in: path
            required: true
            schema:
              type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string

components:
  schemas:
    CarDto:
      required:
        - id
        - brand
        - model
        - owner
        - mechanic
        - color
        - tuition
      type: object
      properties:
        id:
          type: integer
          format: int64
        brand:
          type: string
        model:
          type: string
        owner:
          type: string
        mechanic:
          type: string
        workshop:
          type: integer
          format: int64
        color:
          type: string
        tuition:
          type: string

    CarCreateDto:
      required:
        - brand
        - model
        - owner
        - mechanic
        - workshop
        - color
        - tuition
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        owner:
          type: string
        mechanic:
          type: string
        workshop:
          type: integer
          format: int64
        color:
          type: string
        tuition:
          type: string

    CarReadDto:
      required:
        - id
        - brand
        - model
        - owner
        - mechanic
        - workshop
        - color
        - tuition
      type: object
      properties:
        id:
          type: integer
          format: int64
        brand:
          type: string
        model:
          type: string
        owner:
          type: string
        mechanic:
          type: string
        workshop:
          type: integer
          format: int64
        color:
          type: string
        tuition:
          type: string

    CarUpdateDto:
      required:
        - brand
        - model
        - owner
        - mechanic
        - workshop
        - color
        - tuition
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        owner:
          type: string
        mechanic:
          type: string
        workshop:
          type: integer
          format: int64
        color:
          type: string
        tuition:
          type: string

    CarPatchDto:
      required:
        - brand
        - model
        - owner
        - mechanic
        - workshop
        - color
        - tuition
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        mechanic:
          type: string
        workshop:
          type: integer
          format: int64
        color:
          type: string
        tuition:
          type: string

    ClientDto:
      required:
        - Dni
        - name
        - lastName
        - phone
        - car
      type: object
      properties:
        Dni:
          type: string
        name:
          type: string
        lastName:
          type: string
        phone:
          type: integer
          format: int64
        car:
          type: integer
          format: int64

    ClientCreateDto:
      required:
        - name
        - lastName
        - phone
        - car
      type: object
      properties:
        name:
          type: string
        lastName:
          type: string
        phone:
          type: integer
          format: int64
        car:
          type: integer
          format: int64

    ClientReadDto:
      required:
        - Dni
        - name
        - lastName
        - phone
        - car
      type: object
      properties:
        Dni:
          type: string
        name:
          type: string
        lastName:
          type: string
        phone:
          type: integer
          format: int64
        car:
          type: integer
          format: int64

    ClientUpdateDto:
      required:
        - name
        - lastName
        - phone
        - car
      type: object
      properties:
        name:
          type: string
        lastName:
          type: string
        phone:
          type: integer
          format: int64
        car:
          type: integer
          format: int64

    ClientPatchDto:
      required:
        - name
        - lastName
        - phone
        - car
      type: object
      properties:
        name:
          type: string
        lastName:
          type: string
        phone:
          type: integer
          format: int64
        car:
          type: integer
          format: int64

    MechanicDto:
      required:
        - Dni
        - name
        - lastName
        - phone
        - workshop
        - car
      type: object
      properties:
        Dni:
          type: string
        name:
          type: string
        lastName:
          type: string
        phone:
          type: integer
          format: int64
        workshop:
          type: integer
          format: int64
        car:
          type: integer
          format: int64

    MechanicCreateDto:
      required:
        - name
        - lastName
        - phone
        - workshop
        - car
      type: object
      properties:
        name:
          type: string
        lastName:
          type: string
        phone:
          type: integer
          format: int64
        workshop:
          type: integer
          format: int64
        car:
          type: integer
          format: int64

    MechanicReadDto:
      required:
        - Dni
        - name
        - lastName
        - phone
        - workshop
        - car

      type: object
      properties:
        Dni:
          type: string
        name:
          type: string
        lastName:
          type: string
        phone:
          type: integer
          format: int64
        workshop:
          type: integer
          format: int64
        car:
          type: integer
          format: int64

    MechanicUpdateDto:
      required:
        - name
        - lastName
        - phone
        - workshop
        - car
      type: object
      properties:
        name:
          type: string
        lastName:
          type: string
        phone:
          type: integer
          format: int64
        workshop:
          type: integer
          format: int64
        car:
          type: integer
          format: int64