# src/main/java/es/taller/alex/controller/dtos/mechaic/MechanicReadDto.java
package es.taller.alex.controller.dtos.mechaic;

import lombok.Data;

@Data
public class MechanicReadDto {

    private String Dni;

    private String name;

    private String lastName;

    private Long phone;

    private Long workshop;

    private Long car;
}